# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pokemon API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/api/pokemon/{term}": {
            "get": {
                "tags": [
                    "Pokemon"
                ],
                "summary": "Buscar pokémon por nombre",
                "operationId": "8c2e829d98b957cd4d9c91cac6b67450",
                "parameters": [
                    {
                        "name": "term",
                        "in": "path",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "char"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pokémon que coinciden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "charmander"
                                            },
                                            "url": {
                                                "type": "string",
                                                "example": "https://pokeapi.co/api/v2/pokemon/4/"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al consultar PokeAPI",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "PokeAPI error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pokemon/details/{name}": {
            "get": {
                "tags": [
                    "Pokemon"
                ],
                "summary": "Obtener datos de un pokémon por nombre",
                "operationId": "6c5f4ff247b8c5432fec83a18b6548e4",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Nombre del pokémon",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pikachu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos completos del pokémon (respuesta directa de PokeAPI)"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Pokemon",
            "description": "Pokemon"
        }
    ]
}